app/requirements.txt
mccabe==0.6.1
oauth2client==4.1.3
oauthlib==3.0.1
redis==3.2.0
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3>=1.24.1
beautifulsoup4==4.8.0
soupsieve>=1.2
scraperapi-sdk==0.2.2
gunicorn==20.0.4
Flask==1.1.1
